// js_modules/dagit/src/types/TypeSchemaFragment.ts
export type TypeSchemaFragment_RegularType_innerTypes = TypeSchemaFragment_RegularType_innerTypes_RegularType | TypeSchemaFragment_RegularType_innerTypes_CompositeType;

export interface TypeSchemaFragment_RegularType_typeAttributes {
  isNamed: boolean;
}

export interface TypeSchemaFragment_RegularType {
  name: string;
  description: string | null;
  isDict: boolean;
  isList: boolean;
  isNullable: boolean;
  innerTypes: TypeSchemaFragment_RegularType_innerTypes[];
  typeAttributes: TypeSchemaFragment_RegularType_typeAttributes;
}
// ...
export type TypeSchemaFragment_CompositeType_innerTypes = TypeSchemaFragment_CompositeType_innerTypes_RegularType | TypeSchemaFragment_CompositeType_innerTypes_CompositeType;

export interface TypeSchemaFragment_CompositeType_typeAttributes {
  isNamed: boolean;
}

export interface TypeSchemaFragment_CompositeType_fields_type {
  name: string;
}

export interface TypeSchemaFragment_CompositeType_fields {
  name: string;
  description: string | null;
  type: TypeSchemaFragment_CompositeType_fields_type;
  isOptional: boolean;
}

export interface TypeSchemaFragment_CompositeType {
  name: string;
  description: string | null;
  isDict: boolean;
  isList: boolean;
  isNullable: boolean;
  innerTypes: TypeSchemaFragment_CompositeType_innerTypes[];
  typeAttributes: TypeSchemaFragment_CompositeType_typeAttributes;
  fields: TypeSchemaFragment_CompositeType_fields[];
}

export type TypeSchemaFragment = TypeSchemaFragment_RegularType | TypeSchemaFragment_CompositeType;